requirements.txt
stremlit
google-generativeai